Bypass the checks page for all PR checks, not just for Travis and Azure Pipelines
